refactor(exchanges): minor code refactor
Replace `Function<*, Void>` -> `Consumer<*>`
Move `PingUtil` to util package
