+++ data/src/main/java/com/hxl/data/util/PingUtil.java
@@ -1,4 +1,4 @@
-package com.hxl.data.repository;
+package com.hxl.data.util;
 
 import java.io.IOException;
